split icon picker components into their own files
