+++ components/IconPicker/icon-picker-select-control.js
@@ -1,0 +1,3 @@
+export const IconPickerSelectControl = () => {
+	return null;
+};
