+++ components/IconPicker/index.js
@@ -1,103 +1,5 @@
-/* eslint-disable react/jsx-props-no-spreading */
-import { __ } from '@wordpress/i18n';
-import {
-	Spinner,
-	CheckboxControl,
-	TextControl,
-	BaseControl,
-	NavigableMenu,
-	VisuallyHidden,
-	Dropdown,
-	ToolbarButton,
-} from '@wordpress/components';
-import { useInstanceId } from '@wordpress/compose';
-import { RawHTML, useState } from '@wordpress/element';
-import { useIcons, useIcon } from '../../hooks/use-icons';
-import { useFilteredList } from '../../hooks/use-filtered-list';
-
-export const IconPicker = (props) => {
-	const { value, onChange, label = null } = props;
-
-	const icons = useIcons();
-
-	const instanceId = useInstanceId(IconPicker);
-	const id = `icon-picker-${instanceId}`;
-
-	const [searchTerm, setSearchTerm] = useState('');
-	const [filteredIcons] = useFilteredList(icons, searchTerm);
-
-	return (
-		<BaseControl label={label} id={id} className="component-icon-picker">
-			<TextControl
-				label={<VisuallyHidden>{__('Search for an Icon')}</VisuallyHidden>}
-				onChange={setSearchTerm}
-				value={searchTerm}
-				placeholder={__('Search for an Icon')}
-				className="component-icon-picker__search"
-			/>
-			<NavigableMenu
-				orientation="vertical"
-				className="component-icon-picker__list"
-				style={{
-					display: 'grid',
-					gridTemplateColumns: 'repeat(4, 1fr)',
-					gap: '12px',
-				}}
-			>
-				{filteredIcons.map((icon) => {
-					const isChecked = value?.name === icon.name && value?.iconSet === icon.iconSet;
-					const Label = () => (
-						<span>
-							<Icon key={icon.name} name={icon.name} iconSet={icon.iconSet} />
-							<VisuallyHidden>{icon.label}</VisuallyHidden>
-						</span>
-					);
-					return (
-						<CheckboxControl
-							key={icon.name}
-							label={<Label />}
-							checked={isChecked}
-							onChange={() => onChange(icon)}
-							className="component-icon-picker__checkbox-control"
-						/>
-					);
-				})}
-			</NavigableMenu>
-		</BaseControl>
-	);
-};
-
-export const Icon = (props) => {
-	const { name, iconSet } = props;
-	const icon = useIcon(iconSet, name);
-
-	if (!icon) {
-		return <Spinner />;
-	}
-
-	return <RawHTML>{icon.source}</RawHTML>;
-};
-
-export const IconPickerToolbarButton = (props) => {
-	return (
-		<Dropdown
-			className="component-icom-picker-toolbar-button"
-			contentClassName="component-icom-picker-toolbar-button__content"
-			position="bottom right"
-			renderToggle={({ isOpen, onToggle }) => (
-				<ToolbarButton onClick={onToggle} aria-expanded={isOpen}>
-					{props?.buttonLabel ?? __('Select Icon')}
-				</ToolbarButton>
-			)}
-			renderContent={() => <IconPicker {...props} />}
-		/>
-	);
-};
-
-export const InlineIconPicker = () => {
-	return null;
-};
-
-export const IconPickerSelectControl = () => {
-	return null;
-};
+export { Icon } from './icon';
+export { IconPicker } from './icon-picker';
+export { IconPickerToolbarButton } from './icon-picker-toolbar-button';
+export { InlineIconPicker } from './inline-icon-picker';
+export { IconPickerSelectControl } from './icon-picker-select-control';
