+++ components/IconPicker/inline-icon-picker.js
@@ -1,0 +1,3 @@
+export const InlineIconPicker = () => {
+	return null;
+};
